Renamed Matrix.py to CreateTableMatrix.py and added more details to metadata.
